Refactor row value evaluation, refactor candidate set update
Refactored row value to use a sliding window algorithm instead of the previous string based pattern matching. The sliding window has a size of 5, and it counts the amount of player and AI markers within the window. If there's only either player or AI markers in the current window, the value of the found pattern is added to the current row score.

Candidate set is now updated in place instead of constantly creating copies.