Split minimax logic, fix TT in min, prioritize moves that cause cutoff
The minimax logic was split into one clear conditional that checks if maximizing or not.

Fixed the transposition table to store lower bound flag only if value >= beta_orig.

Added a history table for storing moves that caused an alpha-beta cutoff. The value of the move in the history table increases the more cutoffs it causes.

Move ordering is now based on
  1) best previous move
  2) historical value of moves (i.e. amount of cutoffs caused)
  3) Chebyshev distance from previous move