Add user guide, add toggle for who starts game, update testing report
Added the user guide document. It has some basic instructions for installation and gameplay.

There is now a bool toggle that determines who starts the game. If AI starts the game, they always play on (9,9). Marker colour is now based on who starts and no longer hardcoded.

Updated the testing report with current coverage and test listing with latest additions.